RequestHeader method added, code refactor

+++ src/main/java/ro/fasttrackit/homework5/repository/CountryRepository.java
@@ -3,5 +3,5 @@ package ro.fasttrackit.homework5.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ro.fasttrackit.homework5.domain.Country;
 
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
 }
